pyproject.toml
[tool.poetry]
name = "pykeycloakjwt"
version = "0.1.0"
description = ""
authors = ["Pengfei Liu <liu.pengfei@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
PyJWT = {extras = ["crypto"], version = "^2.4.0"}
python-keycloak = "^0.27.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
pytest = "^7.1.2"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
